pyproject.toml
[tool.poetry]
name = "rishat"
version = "0.1.0"
description = ""
authors = ["subbotinav"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
stripe = "^5.1.1"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.1"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
